First pass at a sane vim config

+++ vim/update_bundles.rb
@@ -4,6 +4,7 @@ git_bundles = [
   "git://github.com/astashov/vim-ruby-debugger.git",
   "git://github.com/msanders/snipmate.vim.git",
   "git://github.com/scrooloose/nerdtree.git",
+  "git//github.com/scrooloose/nerdcommenter.git",
   "git://github.com/timcharper/textile.vim.git",
   "git://github.com/tpope/vim-cucumber.git",
   "git://github.com/tpope/vim-fugitive.git",
@@ -18,6 +19,14 @@ git_bundles = [
   "git://github.com/tsaleh/vim-supertab.git",
   "git://github.com/tsaleh/vim-tcomment.git",
   "git://github.com/vim-ruby/vim-ruby.git",
+  "git://github.com/mileszs/ack.vim.git",
+  "git://github.com/tpope/vim-vividchalk.git",
+  "git://github.com/taq/vim-rspec.git",
+  "git://github.com/Townk/vim-autoclose.git",
+  "git://github.com/scrooloose/nerdcommenter.git",
+  "git://github.com/robgleeson/vim-markdown-preview.git",
+  "git://github.com/pangloss/vim-javascript.git",
+  "git://github.com/michaeljsmith/vim-indent-object.git"
 ]
 
 
@@ -25,6 +34,9 @@ vim_org_scripts = [
   ["IndexedSearch", "7062",  "plugin"],
   ["gist",          "12732", "plugin"],
   ["jquery",        "12107", "syntax"],
+  ["bufexplorer",   "12904", "zip"],
+  ["taglist",       "7701",  "zip"],
+  ["color-sampler", "12179", "zip"]
 ]
 
 require 'fileutils'
@@ -34,22 +46,37 @@ bundles_dir = File.join(File.dirname(__FILE__), "bundle")
 
 FileUtils.cd(bundles_dir)
 
-puts "Trashing everything (lookout!)"
-Dir["*"].each {|d| FileUtils.rm_rf d }
+notrash = ARGV.include?('--notrash')
+
+unless notrash
+  puts "Trashing everything (lookout!)"
+  Dir["*"].each {|d| FileUtils.rm_rf d }
+end
 
 git_bundles.each do |url|
   dir = url.split('/').last.sub(/\.git$/, '')
+  if notrash && File.exists?(dir)
+    puts "  Skipping #{dir}"
+    next
+  end
   puts "  Unpacking #{url} into #{dir}"
   `git clone #{url} #{dir}`
   FileUtils.rm_rf(File.join(dir, ".git"))
 end
 
 vim_org_scripts.each do |name, script_id, script_type|
-  puts "  Downloading #{name}"
-  local_file = File.join(name, script_type, "#{name}.vim")
+  local_file = File.join(name, script_type, "#{name}.#{script_type == 'zip' ? 'zip' : 'vim'}")
+  if notrash && File.exists?(local_file)
+    puts "  Skipping #{local_file}"
+    next
+  end
+  puts " Downloading #{name}"
   FileUtils.mkdir_p(File.dirname(local_file))
   File.open(local_file, "w") do |file|
     file << open("http://www.vim.org/scripts/download_script.php?src_id=#{script_id}").read
+  end
+  if script_type == 'zip'
+    %x(unzip -d #{name} #{local_file})
   end
 end
